add algotitms

+++ 1002/1002.c
@@ -1,16 +1,18 @@
 #include <stdio.h>
 
 int main(){
-    //Constante pi
-    float pi = 3.14159;
-    //Variável referente ao raio   
-    float raio;
+
+    //Define a Constante pi
+    double pi = 3.14159;
+
+    //Define a Variável referente ao raio   
+    double raio;
    
-    //Entrada do raio
-    scanf("%f", &raio);
+    //Entrada do valor do raio
+    scanf("%lf", &raio);
    
     //Calcula a área do círculo
-    float area = pi * (raio * raio);
+    double area = pi * (raio * raio);
    
     //Imprime a área com 4 casas decimais depois da vírgula
     printf("A=%.4f\n", area);
